feat: change General to Public
