+++ web-client/src/ducks/index.ts
@@ -1,8 +1,8 @@
 import auth from './auth/reducer';
 import findRequestReducer from './findRequests/reducer';
-import myOfferReducer from './GeneralOffers/reducer';
-import myRequestReducer from './GeneralRequests/reducer';
 import profile from './profile/reducer';
+import myOfferReducer from './PublicOffers/reducer';
+import myRequestReducer from './PublicRequests/reducer';
 import search from './search/reducer';
 import offers from './specificOffers/reducer';
 import timeline from './timeline/reducer';
