+++ web-client/src/modules/timeline/containers/TimelineViewContainer/TimelineViewContainer.tsx
@@ -4,8 +4,8 @@ import { useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 import { resetSetRequestState } from 'src/ducks/findRequests/actions';
-import { RequestState } from 'src/ducks/GeneralRequests/types';
 import { ProfileState } from 'src/ducks/profile/types';
+import { RequestState } from 'src/ducks/PublicRequests/types';
 import { OffersState } from 'src/ducks/specificOffers/types';
 import { getPostWithOffersAndTimelineItems } from 'src/ducks/timeline/functions';
 import { firestore as firestore2 } from 'src/firebase';
